Move options and extensions
