+++ source/dotnet/wholesale-api/WebApi/Extensions/DependencyInjection/EventsExtensions.cs
@@ -23,7 +23,7 @@ using Energinet.DataHub.Wholesale.Calculations.Interfaces.GridArea;
 using Energinet.DataHub.Wholesale.Events.Application.Workers;
 using Google.Protobuf.Reflection;
 
-namespace Energinet.DataHub.Wholesale.WebApi.Configuration.Modules;
+namespace Energinet.DataHub.Wholesale.WebApi.Extensions.DependencyInjection;
 
 /// <summary>
 /// Registration of services required for the Events module.
